requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
blockdiag==1.5.4
coverage==4.5.4
elasticsearch==7.0.2
entrypoints==0.3
evtxtoelk==1.0.2
flake8==3.7.8
funcparserlib==0.3.6
hexdump==3.3
importlib-metadata==0.19
mccabe==0.6.1
more-itertools==7.2.0
mypy==0.720
mypy-extensions==0.4.1
numpy==1.17.0
packaging==19.1
pandas==0.25.0
Pillow==6.1.0
pluggy==0.12.0
pprint==0.1
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.2
pytest==5.0.1
pytest-cov==2.7.1
python-dateutil==2.8.0
python-evtx==0.6.1
pytz==2019.2
seqdiag==0.9.6
six==1.12.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.3
wcwidth==0.1.7
webcolors==1.9.1
xmltodict==0.12.0
zipp==0.5.2
